[#54] ViewModelからModelの処理を別で切り出した
